// src/pages/park-center/carManage/data.d.ts
export const OilEnum = {
  NORMAL: {
    text: '燃油',
  },
  BAND: {
    text: '新能源',
  },
};

export const CarTypeEnum = {
  1: {
    text: '小型车',
  },
  2: {
    text: '大型车',
  },
};

export const PaidPackageEnum = {
  NORMAL: {
    text: '产权车',
  },
  BAND: {
    text: '月租车',
  },
  BAND2: {
    text: '储蓄卡',
  },
};

// 套餐用途
export const PackageUseEnum = {
  1: {
    text: '产权',
  },
  2: {
    text: '月租',
  },
  3: {
    text: '免费',
  },
  4: {
    text: '其它',
  },
};
// 车场品牌
export const YardBrandEnum = {
  1: {
    text: '艾科',
  },
  2: {
    text: '富士',
  },
  3: {
    text: '捷顺',
  },
  4: {
    text: '狄耐克',
  },
  5: {
    text: '红门',
  },
};

export const OperationType = {
  1: {
    text: '新增',
  },
  2: {
    text: '删除',
  },
  3: {
    text: '编辑',
  },
  4: {
    text: '续费',
  },
  5: {
    text: '延期',
  },
  6: {
    text: '取消授权',
  },
};

export const BusinessType = {
  1: {
    text: '访客',
  },
  2: {
    text: '黑名单',
  },
  3: {
    text: '车主车辆信息',
  },
  4: {
    text: '车辆办理',
  },
};

export const AuthOpertionSattus = {
  1: {
    text: '下发成功',
  },
  2: {
    text: '无需下发',
  },
  3: {
    text: '下发失败',
  },
};